Add more tests for instructions and move Pulse tests into a new pulse_lib testing file
