+++ test/python/pulse/test_instructions.py
@@ -14,9 +14,40 @@
 
 """Unit tests for pulse instructions."""
 
-from qiskit.pulse import DriveChannel, pulse_lib
-from qiskit.pulse import Delay, Play, ShiftPhase, Snapshot
+from qiskit.pulse import DriveChannel, AcquireChannel, MemorySlot, pulse_lib
+from qiskit.pulse import Delay, Play, ShiftPhase, Snapshot, SetFrequency, Acquire
+from qiskit.pulse.configuration import Kernel, Discriminator
 from qiskit.test import QiskitTestCase
+
+
+class TestAcquire(QiskitTestCase):
+    """Acquisition tests."""
+
+    def test_can_construct_valid_acquire_command(self):
+        """Test if valid acquire command can be constructed."""
+        kernel_opts = {
+            'start_window': 0,
+            'stop_window': 10
+        }
+        kernel = Kernel(name='boxcar', **kernel_opts)
+
+        discriminator_opts = {
+            'neighborhoods': [{'qubits': 1, 'channels': 1}],
+            'cal': 'coloring',
+            'resample': False
+        }
+        discriminator = Discriminator(name='linear_discriminator', **discriminator_opts)
+
+        acq = Acquire(10, AcquireChannel(0), MemorySlot(0),
+                      kernel=kernel, discriminator=discriminator)
+
+        self.assertEqual(acq.duration, 10)
+        self.assertEqual(acq.operands, (10, AcquireChannel(0), MemorySlot(0)))
+        self.assertEqual(acq.discriminator.name, 'linear_discriminator')
+        self.assertEqual(acq.discriminator.params, discriminator_opts)
+        self.assertEqual(acq.kernel.name, 'boxcar')
+        self.assertEqual(acq.kernel.params, kernel_opts)
+        self.assertTrue(acq.name.startswith('acquire'))
 
 
 class TestDelay(QiskitTestCase):
@@ -29,6 +60,25 @@ class TestDelay(QiskitTestCase):
         self.assertEqual(delay.name, "test_name")
         self.assertEqual(delay.duration, 10)
         self.assertEqual(delay.operands, (10, DriveChannel(0)))
+        self.assertEqual(delay, Delay(10, DriveChannel(0)))
+        self.assertNotEqual(delay, Delay(11, DriveChannel(1)))
+        self.assertEqual(repr(delay), "Delay(10, DriveChannel(0), name='test_name')")
+
+
+class TestSetFrequency(QiskitTestCase):
+    """Set frequency tests."""
+
+    def test_freq(self):
+        """Test set frequency basic functionality."""
+        set_freq = SetFrequency(4.5e9, DriveChannel(1), name='test')
+
+        self.assertEqual(set_freq.frequency, 4.5e9)
+        self.assertEqual(set_freq.duration, 0)
+        self.assertEqual(set_freq.operands, (4.5e9, DriveChannel(1)))
+        self.assertEqual(set_freq, SetFrequency(4.5e9, DriveChannel(1), name='test'))
+        self.assertNotEqual(set_freq, SetFrequency(4.5e8, DriveChannel(1), name='test'))
+        self.assertEqual(repr(set_freq),
+                         "SetFrequency(4500000000.0, DriveChannel(1), name='test')")
 
 
 class TestShiftPhase(QiskitTestCase):
@@ -41,6 +91,10 @@ class TestShiftPhase(QiskitTestCase):
         self.assertEqual(shift_phase.phase, 1.57)
         self.assertEqual(shift_phase.duration, 0)
         self.assertTrue(shift_phase.name.startswith('shiftphase'))
+        self.assertEqual(shift_phase.operands, (1.57, DriveChannel(0)))
+        self.assertEqual(shift_phase, ShiftPhase(1.57, DriveChannel(0), name='test'))
+        self.assertNotEqual(shift_phase, ShiftPhase(1.57j, DriveChannel(0), name='test'))
+        self.assertEqual(repr(shift_phase)[:-7], "ShiftPhase(1.57, DriveChannel(0), name='shiftphase")
 
 
 class TestSnapshot(QiskitTestCase):
@@ -53,6 +107,8 @@ class TestSnapshot(QiskitTestCase):
         self.assertEqual(snapshot.name, "test_name")
         self.assertEqual(snapshot.type, "state")
         self.assertEqual(snapshot.duration, 0)
+        self.assertNotEqual(snapshot, Delay(10, DriveChannel(0)))
+        self.assertEqual(repr(snapshot), "Snapshot(test_name, state, name='test_name')")
 
 
 class TestPlay(QiskitTestCase):
@@ -60,8 +116,16 @@ class TestPlay(QiskitTestCase):
 
     def test_play(self):
         """Test basic play instruction."""
-        duration = 64
+        duration = 4
         pulse = pulse_lib.SamplePulse([1.0] * duration, name='test')
         play = Play(pulse, DriveChannel(1))
+
         self.assertEqual(play.name, pulse.name)
         self.assertEqual(play.duration, duration)
+        self.assertEqual(repr(play),
+                         "Play(SamplePulse(array([1.+0.j, 1.+0.j, 1.+0.j, 1.+0.j]), name='test'),"
+                         " DriveChannel(1), name='test')")
+
+
+if __name__ == '__main__':
+    unittest.main()
